perf(utils): [utils/format] 优化 createBEM 方法
